# perf(weave): push trace_roots_only pre group by sql filter