--- a/weave/trace_server/calls_query_builder/calls_query_builder.py
+++ b/weave/trace_server/calls_query_builder/calls_query_builder.py
@@ -711,8 +711,8 @@ class CallsQuery(BaseModel):
                 having_conditions_sql, "AND"
             )
 
-        # The op_name, trace_id conditions REQUIRE conditioning on the started_at
-        # field after grouping in the HAVING clause. These filters can remove
+        # The op_name, trace_id, trace_roots conditions REQUIRE conditioning on the
+        # started_at field after grouping in the HAVING clause. These filters remove
         # call starts before grouping, creating orphan call ends. By conditioning
         # on `NOT any(started_at) is NULL`, we filter out orphaned call ends, ensuring
         # all rows returned at least have a call start.
@@ -722,6 +722,11 @@ class CallsQuery(BaseModel):
             table_alias,
         )
         trace_id_sql = process_trace_id_filter_to_conditions(
+            self.hardcoded_filter,
+            pb,
+            table_alias,
+        )
+        trace_roots_only_sql = process_trace_roots_only_filter_to_sql(
             self.hardcoded_filter,
             pb,
             table_alias,
@@ -815,6 +820,7 @@ class CallsQuery(BaseModel):
         {id_mask_sql}
         {id_subquery_sql}
         {sortable_datetime_sql}
+        {trace_roots_only_sql}
         {op_name_sql}
         {trace_id_sql}
         {str_filter_opt_sql}
@@ -1159,6 +1165,26 @@ def process_trace_id_filter_to_conditions(
     return f" AND ({trace_cond} OR {trace_id_field_sql} IS NULL)"
 
 
+def process_trace_roots_only_filter_to_sql(
+    hardcoded_filter: Optional[HardCodedFilter],
+    param_builder: ParamBuilder,
+    table_alias: str,
+) -> str:
+    """Pulls out the trace_roots_only and returns a sql string if there are any trace_roots_only."""
+    if hardcoded_filter is None or not hardcoded_filter.filter.trace_roots_only:
+        return ""
+
+    parent_id_field = get_field_by_name("parent_id")
+    if not isinstance(parent_id_field, CallsMergedAggField):
+        raise TypeError("parent_id is not an aggregate field")
+
+    parent_id_field_sql = parent_id_field.as_sql(
+        param_builder, table_alias, use_agg_fn=False
+    )
+
+    return f"{parent_id_field_sql} IS NULL"
+
+
 def process_calls_filter_to_conditions(
     filter: tsi.CallsFilter,
     param_builder: ParamBuilder,
@@ -1192,11 +1218,6 @@ def process_calls_filter_to_conditions(
         assert_parameter_length_less_than_max("call_ids", len(filter.call_ids))
         conditions.append(
             f"{get_field_by_name('id').as_sql(param_builder, table_alias)} IN {param_slot(param_builder.add_param(filter.call_ids), 'Array(String)')}"
-        )
-
-    if filter.trace_roots_only:
-        conditions.append(
-            f"{get_field_by_name('parent_id').as_sql(param_builder, table_alias)} IS NULL"
         )
 
     if filter.wb_user_ids: