## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

[WB-24605](https://wandb.atlassian.net/browse/WB-24605)

`parent_id` is guaranteed to exist on call starts. This means we can apply the same kinds of optimization as `op_name` and `trace_id` to the `trace_roots_only` hardcoded filter. A ton of queries use this filter (most UI loads) so small improvements here have an outsized impact.

This is a separate optimization as `parent_id = <something>` because in that case we need to also account for if the `parent_id IS NULL`. In _this_ case, that is the entire condition! So we can filter out any call starts that have a parent id.

```sql
SELECT calls_merged.id AS id
FROM calls_merged
WHERE calls_merged.project_id = 'UHJvamVjdEludGVybmFsSWQ6NDA5MzI1OTI='
    AND calls_merged.parent_id IS NULL
    ----------------------------------
GROUP BY (calls_merged.project_id, calls_merged.id)
HAVING (any(calls_merged.deleted_at) IS NULL) 
    AND (NOT (any(calls_merged.started_at) IS NULL)) 
```

## Testing

- calls query unit tests
- client end to end tests
- manual tests in prod clickhouse:
For most queries:
![image](https://github.com/user-attachments/assets/6292dd03-03a6-4c0f-b083-bd454f46eced)

In projects that have lots of nested traces, this becomes a much better optimization. This is the light filter portion of a query in a very large project. The filter is on only root traces, so this optimization massively reduces the number of traces that hit the light filter group by. 

| Environment | Elapsed Time | Rows Read     | Data Read | Peak Memory |
|-------------|--------------|---------------|-----------|-------------|
| Prod        | 4.573s       | 71,858,518     | 10.30 GB  | 23.76 GiB   |
| Branch      | 4.460s       | 71,857,876     | 10.30 GB  | 1.53 GiB    |

Even stats queries are improved:
| Environment | Elapsed Time | Rows Read     | Data Read | Peak Memory   |
|-------------|--------------|---------------|-----------|----------------|
| Prod        | 3.205s       | 12,232,893     | 1.76 GB   | 4.73 GiB       |
| Branch      | 2.593s       | 12,236,307     | 1.76 GB   | 277.87 MiB     | 

Weekly traffic on one query that will be affected:
| Env | Runs | Errors | Avg. Read Rows | Avg. Memory Usage | p50 | p90 | p99 |
|-------------|------------------|--------|----------------|-------------------|-------------|-------------|-------------|
| Prod        | 15,623           | 0      | 38,662.03      | 111.77 MiB        | 0.05s       | 0.321s      | 2.042s      |

[WB-24605]: https://wandb.atlassian.net/browse/WB-24605?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ